refactor: refactor buttons
