--- src/app/cryptocurrencies/Table/Dialogs/styles.ts
@@ -1,10 +1,0 @@
-import PrimaryButton from '@/components/PrimaryButton';
-import styled from 'styled-components';
-
-export const StyledDangerButton = styled(PrimaryButton)`
-	background-color: #ff1b1b;
-
-	&:hover {
-		background-color: #f45050;
-	}
-`;
