+++ src/components/Button/PrimaryButton/index.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+import BaseButton, { type BaseButtonProps } from '../BaseButton';
+
+export type PrimaryButtonProps = BaseButtonProps;
+
+export const PrimaryButton = styled(BaseButton)`
+	background-color: #007690;
+
+	&:hover {
+		background-color: #2e899d;
+	}
+`;
+
+export default PrimaryButton;
