--- src/components/PrimaryButton/index.tsx
@@ -1,14 +1,0 @@
-import { StyledButton } from './styles';
-
-type PrimaryButtonProps = React.ButtonHTMLAttributes<HTMLButtonElement>;
-
-export default function PrimaryButton({
-	children,
-	...buttonProps
-}: React.PropsWithChildren<PrimaryButtonProps>) {
-	return (
-		<>
-			<StyledButton {...buttonProps}>{children}</StyledButton>
-		</>
-	);
-}
